MONIT-26425 - Support OTLP Cumulative Histograms (#8)
* Handle Cumulative Histogram
* Pull out OtlpProtobufPointUtils into separate test class
