+++ proxy/src/main/java/com/wavefront/agent/listeners/otlp/OtlpProtobufPointUtils.java
@@ -2,28 +2,13 @@ package com.wavefront.agent.listeners.otlp;
 
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.collect.Lists;
+
 import com.wavefront.agent.handlers.ReportableEntityHandler;
 import com.wavefront.agent.preprocessor.ReportableEntityPreprocessor;
 import com.wavefront.common.MetricConstants;
 import com.wavefront.sdk.common.Pair;
 
-import io.opentelemetry.proto.collector.metrics.v1.ExportMetricsServiceRequest;
-import io.opentelemetry.proto.common.v1.AnyValue;
-import io.opentelemetry.proto.common.v1.KeyValue;
-import io.opentelemetry.proto.metrics.v1.Gauge;
-import io.opentelemetry.proto.metrics.v1.InstrumentationLibraryMetrics;
-import io.opentelemetry.proto.metrics.v1.Metric;
-import io.opentelemetry.proto.metrics.v1.NumberDataPoint;
-import io.opentelemetry.proto.metrics.v1.ResourceMetrics;
-import io.opentelemetry.proto.metrics.v1.Sum;
-import io.opentelemetry.proto.metrics.v1.Summary;
-import io.opentelemetry.proto.metrics.v1.SummaryDataPoint;
-import io.opentelemetry.proto.resource.v1.Resource;
 import org.jetbrains.annotations.NotNull;
-import wavefront.report.Annotation;
-import wavefront.report.ReportPoint;
-
-import javax.annotation.Nullable;
 
 import java.util.ArrayList;
 import java.util.Collection;
@@ -35,6 +20,25 @@ import java.util.function.Supplier;
 import java.util.logging.Logger;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
+
+import javax.annotation.Nullable;
+import io.opentelemetry.proto.collector.metrics.v1.ExportMetricsServiceRequest;
+import io.opentelemetry.proto.common.v1.AnyValue;
+import io.opentelemetry.proto.common.v1.KeyValue;
+import io.opentelemetry.proto.metrics.v1.Gauge;
+import io.opentelemetry.proto.metrics.v1.Histogram;
+import io.opentelemetry.proto.metrics.v1.HistogramDataPoint;
+import io.opentelemetry.proto.metrics.v1.InstrumentationLibraryMetrics;
+import io.opentelemetry.proto.metrics.v1.Metric;
+import io.opentelemetry.proto.metrics.v1.NumberDataPoint;
+import io.opentelemetry.proto.metrics.v1.ResourceMetrics;
+import io.opentelemetry.proto.metrics.v1.Sum;
+import io.opentelemetry.proto.metrics.v1.Summary;
+import io.opentelemetry.proto.metrics.v1.SummaryDataPoint;
+import io.opentelemetry.proto.resource.v1.Resource;
+import wavefront.report.Annotation;
+import wavefront.report.ReportPoint;
+
 
 public class OtlpProtobufPointUtils {
   public final static Logger OTLP_DATA_LOGGER = Logger.getLogger("OTLPDataLogger");
@@ -70,9 +74,9 @@ public class OtlpProtobufPointUtils {
         OTLP_DATA_LOGGER.finest(() -> "Inbound OTLP Instrumentation Library: " +
             instrumentationLibraryMetrics.getInstrumentationLibrary());
         for (Metric otlpMetric : instrumentationLibraryMetrics.getMetricsList()) {
+          OTLP_DATA_LOGGER.finest(() -> "Inbound OTLP Metric: " + otlpMetric);
           List<ReportPoint> points = transform(otlpMetric, sourceAndResourceAttrs._2,
               preprocessor, sourceAndResourceAttrs._1);
-          OTLP_DATA_LOGGER.finest(() -> "Inbound OTLP Metric: " + otlpMetric);
           OTLP_DATA_LOGGER.finest(() -> "Converted Wavefront Metric: " + points);
 
           wfPoints.addAll(points);
@@ -115,6 +119,9 @@ public class OtlpProtobufPointUtils {
       points.addAll(transformSum(otlpMetric.getName(), otlpMetric.getSum(), resourceAttrs));
     } else if (otlpMetric.hasSummary()) {
       points.addAll(transformSummary(otlpMetric.getName(), otlpMetric.getSummary(), resourceAttrs));
+    } else if (otlpMetric.hasHistogram()) {
+      points.addAll(transformHistogram(otlpMetric.getName(), otlpMetric.getHistogram(),
+          resourceAttrs));
     } else {
       throw new IllegalArgumentException("Otel: unsupported metric type for " + otlpMetric.getName());
     }
@@ -162,6 +169,71 @@ public class OtlpProtobufPointUtils {
     return points;
   }
 
+  private static List<ReportPoint> transformHistogram(String name, Histogram histogram, List<KeyValue> resourceAttrs) {
+
+    switch (histogram.getAggregationTemporality()) {
+      case AGGREGATION_TEMPORALITY_CUMULATIVE:
+        return transformCumulativeHistogram(name, histogram, resourceAttrs);
+      default:
+        throw new IllegalArgumentException("OTel: histogram with unsupported aggregation temporality "
+            + histogram.getAggregationTemporality().name());
+    }
+  }
+
+  private static List<ReportPoint> transformCumulativeHistogram(String name, Histogram histogram, List<KeyValue> resourceAttrs) {
+
+    List<ReportPoint> reportPoints = new ArrayList<>();
+    for (HistogramDataPoint point : histogram.getDataPointsList()) {
+      reportPoints.addAll(transformHistogramDataPoint(name, point, resourceAttrs));
+    }
+
+    return reportPoints;
+  }
+
+  private static List<ReportPoint> transformHistogramDataPoint(String name,
+                                                               HistogramDataPoint point, List<KeyValue> resourceAttrs) {
+    if (point.getExplicitBoundsCount() != point.getBucketCountsCount() - 1) {
+      throw new IllegalArgumentException("OTel: histogram " + name + " data point has mismatched " +
+          "explicit bound counts (" + point.getExplicitBoundsCount() + ") and bucket counts (" +
+          (point.getBucketCountsCount() - 1) + ")");
+    }
+
+    List<ReportPoint> reportPoints = new ArrayList<>();
+    int currentIndex = 0;
+    long cumulativeBucketCount = 0;
+    for (; currentIndex < point.getExplicitBoundsCount(); currentIndex++) {
+      cumulativeBucketCount += point.getBucketCounts(currentIndex);
+      // we have to create a new builder every time as the annotations are getting appended after
+      // each iteration
+      ReportPoint rp = pointWithAnnotations(name, point.getAttributesList(), resourceAttrs,
+          point.getTimeUnixNano())
+          .setValue(cumulativeBucketCount)
+          .build();
+      handleDupAnnotation(rp);
+      rp.getAnnotations().put("le", String.valueOf(point.getExplicitBounds(currentIndex)));
+      reportPoints.add(rp);
+    }
+
+    ReportPoint rp = pointWithAnnotations(name, point.getAttributesList(), resourceAttrs,
+        point.getTimeUnixNano())
+        .setValue(cumulativeBucketCount + point.getBucketCounts(currentIndex))
+        .build();
+    handleDupAnnotation(rp);
+    rp.getAnnotations().put("le", "+Inf");
+    reportPoints.add(rp);
+
+    return reportPoints;
+  }
+
+  private static void handleDupAnnotation(ReportPoint rp) {
+    if (rp.getAnnotations().containsKey("le")) {
+      String val = rp.getAnnotations().get("le");
+      rp.getAnnotations().remove("le");
+      rp.getAnnotations().put("_le", val);
+    }
+  }
+
+
   private static Collection<ReportPoint> transformGauge(String name, Gauge gauge,
                                                         List<KeyValue> resourceAttrs) {
     if (gauge.getDataPointsCount() == 0) {
@@ -177,8 +249,7 @@ public class OtlpProtobufPointUtils {
 
   @NotNull
   private static ReportPoint transformNumberDataPoint(String name, NumberDataPoint point, List<KeyValue> resourceAttrs) {
-    return pointWithAnnotations(name, point.getAttributesList(), resourceAttrs)
-        .setTimestamp(TimeUnit.NANOSECONDS.toMillis(point.getTimeUnixNano()))
+    return pointWithAnnotations(name, point.getAttributesList(), resourceAttrs, point.getTimeUnixNano())
         .setValue(point.getAsDouble())
         .build();
   }
@@ -187,12 +258,10 @@ public class OtlpProtobufPointUtils {
   private static List<ReportPoint> transformSummaryDataPoint(String name, SummaryDataPoint point, List<KeyValue> resourceAttrs) {
     List<ReportPoint> toReturn = new ArrayList<>();
     List<KeyValue> pointAttributes = replaceQuantileTag(point.getAttributesList());
-    toReturn.add(pointWithAnnotations(name + "_sum", pointAttributes, resourceAttrs)
-        .setTimestamp(TimeUnit.NANOSECONDS.toMillis(point.getTimeUnixNano()))
+    toReturn.add(pointWithAnnotations(name + "_sum", pointAttributes, resourceAttrs, point.getTimeUnixNano())
         .setValue(point.getSum())
         .build());
-    toReturn.add(pointWithAnnotations(name + "_count", pointAttributes, resourceAttrs)
-        .setTimestamp(TimeUnit.NANOSECONDS.toMillis(point.getTimeUnixNano()))
+    toReturn.add(pointWithAnnotations(name + "_count", pointAttributes, resourceAttrs, point.getTimeUnixNano())
         .setValue(point.getCount())
         .build());
     for (SummaryDataPoint.ValueAtQuantile q : point.getQuantileValuesList()) {
@@ -202,8 +271,7 @@ public class OtlpProtobufPointUtils {
           .setValue(AnyValue.newBuilder().setDoubleValue(q.getQuantile()).build())
           .build();
       attributes.add(quantileTag);
-      toReturn.add(pointWithAnnotations(name, attributes, resourceAttrs)
-          .setTimestamp(TimeUnit.NANOSECONDS.toMillis(point.getTimeUnixNano()))
+      toReturn.add(pointWithAnnotations(name, attributes, resourceAttrs, point.getTimeUnixNano())
           .setValue(q.getValue())
           .build());
     }
@@ -229,7 +297,7 @@ public class OtlpProtobufPointUtils {
   }
 
   @NotNull
-  private static ReportPoint.Builder pointWithAnnotations(String name, List<KeyValue> pointAttributes, List<KeyValue> resourceAttrs) {
+  private static ReportPoint.Builder pointWithAnnotations(String name, List<KeyValue> pointAttributes, List<KeyValue> resourceAttrs, long timeInNs) {
     ReportPoint.Builder builder = ReportPoint.newBuilder().setMetric(name);
     Map<String, String> annotations = new HashMap<>();
     List<KeyValue> otlpAttributes = Stream.of(resourceAttrs, pointAttributes)
@@ -238,6 +306,7 @@ public class OtlpProtobufPointUtils {
       annotations.put(a.getKey(), a.getValue());
     }
     builder.setAnnotations(annotations);
+    builder.setTimestamp(TimeUnit.NANOSECONDS.toMillis(timeInNs));
     return builder;
   }
 }
